Fix gradebook tests
