+++ nbgrader/tests/formgrader/base.py
@@ -55,6 +55,30 @@ class BaseTestFormgrade(object):
         self._wait_for_element("notebook-container")
         self._check_url(url)
 
+    def _wait_for_formgrader(self, url):
+        self._wait_for_element("notebook-container")
+        page_loaded = lambda browser: browser.execute_script(
+            """
+            if (!(typeof MathJax !== "undefined" && MathJax !== undefined && MathJax.loaded)) {
+                return false;
+            }
+            if (!(typeof formgrader !== "undefined" && formgrader !== undefined)) {
+                return false;
+            }
+
+            if (!(formgrader.grades !== undefined && formgrader.grades.loaded)) {
+                return false;
+            }
+
+            if (!(formgrader.comments !== undefined && formgrader.comments.loaded)) {
+                return false;
+            }
+
+            return true;
+            """)
+        WebDriverWait(self.browser, 10).until(page_loaded)
+        self._check_url(url)
+
     def test_start(self):
         # This is just a fake test, since starting up the browser and formgrader
         # can take a little while. So if anything goes wrong there, this test
